requirements.txt
Django==1.10
djangorestframework==3.5
libpq-dev
psycopg2 >= 2.5
django-rest-swagger
markdown
requests